destroy unit in organization structure

+++ app/assets/javascripts/organization_structure_tree_view.js
@@ -30,7 +30,7 @@ OrganizationStructureTreeView.createParentNode = function(parent_id, structure){
   var html = "<ul>";
 
   html += "<li id='node-" + structure.id + "'>";
-  html += "<a href='#' class='highligh-on-hover'>" + structure.unit_id + "</a>";
+  html += "<a href='#' class='highligh-on-hover has-tooltip' title='" + structure.unit_id + structure.name + "' data-original-title='" + structure.unit_id + structure.name + "'>" + structure.unit_id + "</a>";
   html += "</li>";
   html += "</ul>";
 
@@ -46,12 +46,12 @@ OrganizationStructureTreeView.createHTML = function(parent_id, structure){
     if(parent_id == structure[i].parent_id){
 
       html += "<li id='node-" + structure[i].id + "'>";
-      html += "<a href='#' class='highligh-on-hover'>" + structure[i].unit_id + "</a>";
+      html += "<a href='#' class='highligh-on-hover has-tooltip' title='" + structure[i].unit_id + structure[i].name + "' data-original-title='" + structure[i].unit_id + structure[i].name + "'>" + structure[i].unit_id + "</a>";
       html += "<span>";
       html += "<a class='tree-links-no-border' data-remote=true href='/organization/organization_structures/" + structure[i].id + "/edit'>";
       html += "<img src='/assets/edit-icon-16.png' class='icon-list has-tooltip' title='Edit' data-original-title='Edit'>";
       html += "</a>"
-      html += "<a class='tree-links-no-border' data-remote=true data-method='delete' href='/organization/organization_structures/" + structure[i].id + "'>";
+      html += "<a class='tree-links-no-border' data-confirm='Are you sure? Units under selected unit will also be deleted?' data-remote=true data-method='delete' href='/organization/organization_structures/" + structure[i].id + "'>";
       html += "<img src='/assets/delete-icon-16.png' class='icon-list has-tooltip' title='Destroy' data-original-title='Destroy'>";
       html += "</a>"
       html += "</span>";
